refactor: extract out check code from emit
Closes #15535
