+++ cli/main.rs
@@ -73,6 +73,7 @@ use crate::graph_util::graph_valid;
 use crate::proc_state::ProcState;
 use crate::resolver::ImportMapResolver;
 use crate::resolver::JsxResolver;
+use crate::tools::check;
 
 use args::CliOptions;
 use deno_ast::MediaType;
@@ -500,11 +501,11 @@ async fn create_graph_and_maybe_check(
     }
     let maybe_config_specifier = ps.options.maybe_config_file_specifier();
     let cache = TypeCheckCache::new(&ps.dir.type_checking_cache_db_file_path());
-    let check_result = emit::check(
+    let check_result = check::check(
       &graph.roots,
       Arc::new(RwLock::new(graph.as_ref().into())),
       &cache,
-      emit::CheckOptions {
+      check::CheckOptions {
         type_check_mode: ps.options.type_check_mode(),
         debug,
         maybe_config_specifier,
