+++ cli/tools/mod.rs
@@ -1,6 +1,7 @@
 // Copyright 2018-2022 the Deno authors. All rights reserved. MIT license.
 
 pub mod bench;
+pub mod check;
 pub mod coverage;
 pub mod doc;
 pub mod fmt;
